Removed EditButton from OLO Flow and fixed tests and renamed EditLocationModal
