+++ apps/client/src/components/RegisterLookupSummary.test.js
@@ -6,14 +6,12 @@ import addressMockNoMonument from "../__mocks__/addressMockNoMonument";
 import Context from "../__mocks__/context";
 import matchMedia from "../__mocks__/matchMedia";
 import { findTopicBySlug } from "../utils";
-import { EDIT_BUTTON } from "../utils/test-ids";
 import { act, cleanup, fireEvent, render, screen } from "../utils/test-utils";
 import RegisterLookupSummary from "./RegisterLookupSummary";
 
 Object.defineProperty(window, "matchMedia", matchMedia);
 
 jest.mock("react-router-dom", () => ({
-  ...jest.requireActual("react-router-dom"),
   useParams: () => ({ slug: "dakkapel-plaatsen" }),
 }));
 
@@ -46,11 +44,10 @@ describe("RegisterLookupSummary", () => {
 
     // Expect NOT to find zoningplan info
     expect(queryByText("zoningplan")).not.toBeInTheDocument();
-
-    expect(screen.getByText(/wijzig/i)).toBeInTheDocument();
+    expect(screen.queryByText(/wijzig/i)).toBeInTheDocument();
 
     act(() => {
-      fireEvent.click(screen.getByText(/wijzig/i));
+      fireEvent.click(screen.queryByText(/wijzig/i));
     });
   });
 
@@ -89,10 +86,10 @@ describe("RegisterLookupSummary", () => {
     // Expect NOT to find zoningplan info
     expect(queryByText("zoningplan")).not.toBeInTheDocument();
 
-    expect(screen.getByText(/wijzig/i)).toBeInTheDocument();
+    expect(screen.queryByText(/wijzig/i)).toBeInTheDocument();
 
     act(() => {
-      fireEvent.click(screen.getByText(/wijzig/i));
+      fireEvent.click(screen.queryByText(/wijzig/i));
     });
   });
 
@@ -130,12 +127,7 @@ describe("RegisterLookupSummary", () => {
     // Expect to find zoningplan info
     expect(queryByText("zoningplan")).toBeInTheDocument();
 
-    expect(screen.getByText(/wijzig/i)).toBeInTheDocument();
-    expect(screen.queryByTestId(EDIT_BUTTON)).toBeInTheDocument();
-
-    act(() => {
-      fireEvent.click(screen.getByText(/wijzig/i));
-    });
+    expect(screen.queryByText(/wijzig/i)).not.toBeInTheDocument();
   });
 
   it("renders correctly in OLO Flow if building is NOT a monument and a MUNICIPAL cityScape", () => {
@@ -157,12 +149,7 @@ describe("RegisterLookupSummary", () => {
     // Expect to find zoningplan info
     expect(queryByText("zoningplan")).toBeInTheDocument();
 
-    expect(screen.getByText(/wijzig/i)).toBeInTheDocument();
-    expect(screen.queryByTestId(EDIT_BUTTON)).toBeInTheDocument();
-
-    act(() => {
-      fireEvent.click(screen.getByText(/wijzig/i));
-    });
+    expect(screen.queryByText(/wijzig/i)).not.toBeInTheDocument();
   });
 
   it("renders correctly in address from location object", () => {
@@ -181,14 +168,8 @@ describe("RegisterLookupSummary", () => {
     // Expect TO DO find zoningplan info
     expect(queryByText("zoningplan")).toBeInTheDocument();
 
-    expect(screen.getByText(/wijzig/i)).toBeInTheDocument();
-    expect(screen.queryByTestId(EDIT_BUTTON)).toBeInTheDocument();
-
-    act(() => {
-      fireEvent.click(screen.getByText(/wijzig/i));
-    });
+    expect(screen.queryByText(/wijzig/i)).not.toBeInTheDocument();
   });
 
   // @TODO: finish this test when we work on the Location Component
-  // Lines to test: 39,61,69,70,80
 });
